+++ addons/Common/script_component.hpp
@@ -1,4 +1,4 @@
-#define COMPONENT Legacy
+#define COMPONENT Common
 
 #include "\DA\DAC\Common\script_mod.hpp"
 #include "\DA\DAC\Common\script_macros.hpp"
