+++ addons/Common/script_macros.hpp
