+++ addons/Common/script_version.hpp
