--- addons/Core/$PBOPREFIX$
@@ -1,1 +1,0 @@
-DA\DAC\Core
