--- addons/Core/script_component.hpp
@@ -1,4 +1,0 @@
-#define COMPONENT Core
-
-#include "\DA\DAC\Core\script_mod.hpp"
-#include "\DA\DAC\Core\script_macros.hpp"
