+++ addons/Items/config.cpp
@@ -10,7 +10,7 @@ class CfgPatches
         requiredAddons[] =
         {
             "A3_Data_F_Decade_Loadorder",
-            "DAC_Core"
+            "DAC_Common"
         };
         units[] = {};
         weapons[] = {};
