+++ addons/Items/script_component.hpp
@@ -1,4 +1,4 @@
 #define COMPONENT Items
 
-#include "\DA\DAC\Core\script_mod.hpp"
-#include "\DA\DAC\Core\script_macros.hpp"
+#include "\DA\DAC\Common\script_mod.hpp"
+#include "\DA\DAC\Common\script_macros.hpp"
