+++ addons/Objects/script_component.hpp
@@ -1,5 +1,5 @@
 #define COMPONENT Objects
 
-#include "\DA\DAC\Core\script_mod.hpp"
-#include "\DA\DAC\Core\script_macros.hpp"
+#include "\DA\DAC\Common\script_mod.hpp"
+#include "\DA\DAC\Common\script_macros.hpp"
 #include "\DA\DAC\Objects\script_macros.hpp"
