+++ addons/Weapons/script_component.hpp
@@ -1,5 +1,5 @@
 #define COMPONENT Weapons
 #define COMPONENT_BEAUTIFIED Weapons Core
 
-#include "\DA\DAC\Core\script_mod.hpp"
-#include "\DA\DAC\Core\script_macros.hpp"
+#include "\DA\DAC\Common\script_mod.hpp"
+#include "\DA\DAC\Common\script_macros.hpp"
